some updates to OrganizationServiceImplTest
